Move common utils to /common dir
